+++ test/marky-mark.js
@@ -3,7 +3,7 @@
 var markyMark = require('../src/helpers/marky-mark');
 
 exports.testSimpleString = function(test) {
-  test.expect(10);
+  test.expect(12);
 
   var m;
 
@@ -27,6 +27,10 @@ exports.testSimpleString = function(test) {
   m = markyMark('<p><a href="https://youtu.be/VQ01tJ4EWeg?t=2m19s">Dunkirk</a></p>');
   test.ok(m.match(/embed\/VQ01tJ4EWeg/));
   test.ok(!m.match(/t=2m19s/));
+
+  m = markyMark('<p><a href="http://codepen.io/larsenwork/pen/MpjXrb">Codepen</a></p>');
+  test.ok(m.match(/codepen\.io\/larsenwork\/embed\/MpjXrb\//));
+  test.ok(m.match(/<iframe/));
 
   m = markyMark('<p><img src="http://www.example.com" alt="" /></p>');
   test.ok(m.match(/src="http:\/\/www\.example\.com/));